simplified process to create JobManager from dag details + address comments
